[refactor🔨] : front tag 아키텍처 변경
