+++ src/main/java/shop/gaship/gashipfront/tag/adapter/TagAdapterImpl.java
@@ -1,4 +1,4 @@
-package shop.gaship.gashipfront.repository;
+package shop.gaship.gashipfront.tag.adapter;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
@@ -9,7 +9,7 @@ import org.springframework.web.util.UriBuilder;
 import org.springframework.web.util.UriComponentsBuilder;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import shop.gaship.gashipfront.dto.*;
+import shop.gaship.gashipfront.tag.dto.*;
 
 /**
  * packageName    : shop.gaship.gashipfront.repository
@@ -34,7 +34,7 @@ public class TagAdapterImpl implements TagAdapter {
     @Override
     public Mono<Void> addTag(Integer adminId, TagRegisterRequestDto tagRegisterRequestDto) {
         return webClient.post()
-                .uri("/admins/{adminId}/tags",adminId)
+                .uri("/admins/{adminId}/tags", adminId)
                 .contentType(MediaType.APPLICATION_JSON)
                 .bodyValue(tagRegisterRequestDto)
                 .retrieve()
@@ -45,7 +45,7 @@ public class TagAdapterImpl implements TagAdapter {
     @Override
     public Mono<TagResponseDto> findTag(Integer adminId, TagGetRequestDto tagGetRequestDto) {
         return webClient.post()
-                .uri("/admins/{adminId}/tags/{tagId}",adminId,tagGetRequestDto.getTagId())
+                .uri("/admins/{adminId}/tags/{tagId}", adminId, tagGetRequestDto.getTagId())
                 .contentType(MediaType.APPLICATION_JSON)
                 .retrieve()
                 .bodyToMono(TagResponseDto.class);
@@ -54,7 +54,7 @@ public class TagAdapterImpl implements TagAdapter {
     @Override
     public void modifyTag(Integer adminId, TagModifyRequestDto tagModifyRequestDto) {
         webClient.post()
-                .uri("/admins/{adminId}/tags/{tagId}",adminId,tagModifyRequestDto.getTagId())
+                .uri("/admins/{adminId}/tags/{tagId}", adminId, tagModifyRequestDto.getTagId())
                 .contentType(MediaType.APPLICATION_JSON)
                 .bodyValue(tagModifyRequestDto)
                 .retrieve()
@@ -65,7 +65,7 @@ public class TagAdapterImpl implements TagAdapter {
     @Override
     public void removeTag(Integer adminId, TagDeleteRequestDto tagDeleteRequestDto) {
         webClient.post()
-                .uri("/admins/{adminId}/tags/{tagId}",adminId,tagDeleteRequestDto.getTagId())
+                .uri("/admins/{adminId}/tags/{tagId}", adminId, tagDeleteRequestDto.getTagId())
                 .contentType(MediaType.APPLICATION_JSON)
                 .bodyValue(tagDeleteRequestDto)
                 .retrieve()
@@ -78,9 +78,9 @@ public class TagAdapterImpl implements TagAdapter {
         UriBuilder uriBuilder = UriComponentsBuilder.newInstance();
         return webClient.post()
                 .uri(uriBuilder.path("/admins/{adminId}/tags")
-                        .queryParam("page",pageable.getPageNumber())
-                        .queryParam("size",pageable.getPageSize())
-                        .queryParam("sort",pageable.getSort())
+                        .queryParam("page", pageable.getPageNumber())
+                        .queryParam("size", pageable.getPageSize())
+                        .queryParam("sort", pageable.getSort())
                         .build())
                 .contentType(MediaType.APPLICATION_JSON)
                 .retrieve()
