+++ src/main/java/shop/gaship/gashipfront/tag/dto/TagGetRequestDto.java
@@ -1,11 +1,11 @@
-package shop.gaship.gashipfront.dto;
+package shop.gaship.gashipfront.tag.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
 /**
  * packageName    : shop.gaship.gashipfront.dto
- * fileName       : TagRegisterRequestDto
+ * fileName       : TagGetRequestDto
  * author         : choijungwoo
  * date           : 2022/07/15
  * description    :
@@ -15,11 +15,11 @@ import lombok.Getter;
  * 2022/07/15        choijungwoo       최초 생성
  */
 @Getter
-public class TagRegisterRequestDto {
-    private String title;
+public class TagGetRequestDto {
+    private Integer tagId;
 
     @Builder
-    public TagRegisterRequestDto(String title) {
-        this.title = title;
+    public TagGetRequestDto(Integer tagId) {
+        this.tagId = tagId;
     }
 }
