+++ src/main/java/shop/gaship/gashipfront/tag/dto/TagResponseDto.java
@@ -1,11 +1,11 @@
-package shop.gaship.gashipfront.dto;
+package shop.gaship.gashipfront.tag.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
 /**
  * packageName    : shop.gaship.gashipfront.dto
- * fileName       : TagGetRequestDto
+ * fileName       : TagResponseDto
  * author         : choijungwoo
  * date           : 2022/07/15
  * description    :
@@ -15,11 +15,9 @@ import lombok.Getter;
  * 2022/07/15        choijungwoo       최초 생성
  */
 @Getter
-public class TagGetRequestDto {
-    private Integer tagId;
+public class TagResponseDto {
 
     @Builder
-    public TagGetRequestDto(Integer tagId) {
-        this.tagId = tagId;
+    public TagResponseDto() {
     }
 }
